XW-514 | rename CurrentUser to current-user for consistency
