+++ front/scripts/main/current-user.js
